This project use Arduino to control a robotic arm
This project aim to show how to use Arduino to control  a robotic arm to draw a graph based on sensors input. The sensors used in this project includes ultrasonic sensor, button, LDR sensor, Accelerometer. The robotic arm is construct by simple servo mechanismto demonstrate the functinoality of the codes.